chore: refactor theme and router
